Move SystemTestDirectQuery into query package
