+++ java/system-test/system-test-suite/src/main/java/sleeper/systemtest/suite/dsl/query/SystemTestDirectQuery.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package sleeper.systemtest.suite.dsl;
+package sleeper.systemtest.suite.dsl.query;
 
 import sleeper.core.record.Record;
 import sleeper.systemtest.drivers.query.DirectQueryDriver;
